reworked the makefile so that g++ doesn't compile EVERY SINGLE source file every time make is called, reducing build times from approx. 40 seconds to less than 10 seconds

+++ src/basic_components.cpp
@@ -54,7 +54,7 @@ SpriteSheet::~SpriteSheet(){
     SpriteLoader::return_texture(texture);
 }
 
-void SpriteSheet::set_animation_length(unsigned int animationRow, unsigned int length){
+void SpriteSheet::set_animation_length(unsigned int animationRow, unsigned short length){
     numberFramesPerAnimation[animationRow] = length;
 }
